Make end ip address required.
End ip address is not required always, but it automatically sets to start ip address if not set. Manupulation of start ip address in that case gicves errro, as the old end ip address is picked up from state and if it is not valid with respect to the updated start ip address, the update fails.