Fix public ip adress not reflecting in nic ip config on first gyro up
If the nic ip config is primary then then the create was ignored for it as it is created when the nic is created, but any modifications to the attribute 'public-ip-address' is ignored as well.

Solution was to fore a an update on create, which will update the existing primary with the updated values.