Fix rule requiring another gyro up if configured to be disabled during create
A policy rule configured to be disabled during create required another gyro up to actually disable the rule. This was because the the create rule definition stage does not allow setting enabled to false while creating the life cycle policy.

The change, checks for any disabled rule present during creation, and if so fires an update to mark the rule/s to be disabled.