Remove service principal implementation
The current implementation of service principal uses an existing application to create itself, which can cause conflicts depending on the configuration of the application.

We currently do not support managing applications and as such supporting service principal does not help much as for service principals to be used, roles needs to be asigned to them which has to be managed manually.