Move validation check from config load to during create
One of the use case for role assignemnts is for a VM to have an identity being assigned some role on the VM. The VM has an output attribute 'principal-id' which is used.

When creating a VM and role assignment together with role assignment having a reference to the principal id from the vm resource, teh validation fails if the vm is not already created as the id is essentially null.

Moving the validation to the create method solves this, as the refernce is resolved and the value is present by the tim ethe create call for the role assignment is processed.